+++ Duo/Services/CoinsService.cs
@@ -1,6 +1,6 @@
 using System;
-using CourseApp.Repository;
 using CourseApp.ModelViews;
+using Duo.Repositories;
 
 #pragma warning disable IDE0079 // Remove unnecessary suppression
 #pragma warning disable SA1009 // Closing parenthesis should be spaced correctly
